fixed component logic for create goals form
