+++ src/App.js
@@ -24,22 +24,6 @@ function App() {
 
 export default App;
 
-// THIS IS DON'S GOALS FUNCTION WHICH CALLS BasicTextFields which should be in components/creategoals.js
-// NEEDS to be a separate component that can simply be rendered here.
-// export default function Goals() {
-//   return (
-//     <div>
-//       <h2>Create Goals</h2>
-//       <BasicTextFields />
-//       <h2>Frequency</h2>
-//       {/* insert multiple small icons with days of the week for nags.  Use node-con to hook up */}
-//       <h2>Completion Date</h2>
-//       {/* Insert Calendar function for date picker - might need to create a new component */}
-//       <h3>Date Picker</h3>
-//     </div>
-//   );
-// }
-
 // OLD REACT ROUTER ROUTES
 // function App() {
 //   return (
