--- src/components/GoalsIndex.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import TextField from '@material-ui/core/TextField';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
-      width: 200,
-    },
-  },
-}));
-
-export default function BasicTextFields() {
-  const classes = useStyles();
-
-  return (
-    <form className={classes.root} noValidate autoComplete="off">
-      <TextField id="outlined-basic" label= "Goal Title" variant="outlined" />
-      <TextField id="outlined-basic" label="Nag Title" variant="outlined" />
-    </form>
-  );
-}
